requirements.txt
pystray==0.19.5
pillow==10.0.0
pyinstaller==6.1.0
